requirements.txt
beautifulsoup4==4.12.3
certifi==2024.8.30
chardet==5.2.0
charset-normalizer==3.4.0
contourpy==1.3.0
cycler==0.12.1
fonttools==4.54.1
idna==3.10
kiwisolver==1.4.7
matplotlib==3.9.2
numpy==2.1.2
packaging==24.1
pillow==11.0.0
pyparsing==3.2.0
python-dateutil==2.9.0.post0
reportlab==4.2.5
requests==2.32.3
six==1.16.0
soupsieve==2.6
urllib3==2.2.3
word2number==1.1
